+++ client/lib/screens/sessions/session_body.dart
@@ -3,7 +3,8 @@ import 'package:client/screens/sessions/session_drawer_body.dart';
 import 'package:client/screens/sessions/session_operation_bar.dart';
 import 'package:client/screens/sessions/session_sql_editor.dart';
 import 'package:client/screens/sessions/session_sql_results.dart';
-import 'package:client/services/sessions/sessions.dart';
+import 'package:client/services/sessions/session_drawer.dart';
+import 'package:client/services/sessions/session_sql_editor.dart';
 import 'package:client/services/sessions/session_controller.dart';
 import 'package:client/widgets/const.dart';
 import 'package:flutter/material.dart';
@@ -15,7 +16,7 @@ class SessionBodyPage extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    SessionEditorModel sessionEditor = ref.watch(sessionEditorNotifierProvider);
+    SessionEditorModel sessionEditor = ref.watch(sessionEditorNotifierProvider); // todo: 合并或者拆分page, 不要两个页面混在一起刷新
     SessionDrawerModel sessionDrawer = ref.watch(sessionDrawerNotifierProvider);
     SessionController sessionController =
         SessionController.sessionController(sessionDrawer.sessionId);
