+++ client/lib/screens/sessions/session_drawer_body.dart
@@ -2,7 +2,7 @@ import 'package:client/models/sessions.dart';
 import 'package:client/screens/sessions/session_drawer_chat.dart';
 import 'package:client/screens/sessions/session_drawer_metadata.dart';
 import 'package:client/screens/sessions/session_drawer_sql_result.dart';
-import 'package:client/services/sessions/sessions.dart';
+import 'package:client/services/sessions/session_drawer.dart';
 import 'package:client/widgets/const.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
