+++ client/lib/screens/sessions/session_sql_results.dart
@@ -1,5 +1,6 @@
 import 'package:client/models/sessions.dart';
 import 'package:client/services/sessions/session_conn.dart';
+import 'package:client/services/sessions/session_drawer.dart';
 import 'package:client/services/sessions/session_sql_result.dart';
 import 'package:client/services/sessions/sessions.dart';
 import 'package:client/widgets/const.dart';
